requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
channels==4.1.0
Django==4.2.13
djangorestframework==3.15.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.6
sqlparse==0.5.0
typing_extensions==4.12.0
uritemplate==4.1.1
